Expose Auth Token Functions
